+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/item/block/CreativeControllerBlockItem.java
@@ -4,7 +4,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 
-public class CreativeControllerBlockItem extends AbstractNamedBlockItem {
+public class CreativeControllerBlockItem extends NamedBlockItem {
     public CreativeControllerBlockItem(final Block block, final Component name) {
         super(block, new Item.Properties().stacksTo(1), name);
     }