+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -   The upgrade slots now show their supported upgrades.
 -   Different Cable colors. They only connect to same colored cables or the default cable.
+-   Colored variant of exporters. They connect the same way as colored cables.
 -   Support for using the R/U keys in JEI and REI on Grid slots and filtering slots
 -   Crafting Grid.
 -   The crafting matrix in the Crafting Grid now has a button and keybinding to clear to the player inventory.